Fixes #329.
This makes AdminController's `userfileUploadPostHandler` stream its POST data into a file, and makes the HTML form for the userfile upload POST the file directly, instead of wrapping the file in JSON.

Previously, swiftarr.js would create a `FormData()` for this form's values and upload that from the js. This turned the zip file into a uuencoded JSON string. So, now the `fileupload` class special-cases this and makes the form POST the file contents directly in the request body. Part of the reason to do this was to remove the 8-to-6 increase in payload size, part of it was because Vapor's built in content decoders aren't built to decode while streaming in data.

However, this doesn't fully make the operation use streaming. SiteAdminController's `bulkUserfileUploadPostHandler` bounces the call to the API layer by making a client request using Vapor's Client API, wrapped in our own `apiQuery()` call. Vapor's Client API doesn't support chunking the request body. Vapor's Client API is written on top of AsyncHTTPClient, which does support chunking the request body, but moving to that for this one call involves a bunch of work that I don't want to do for a call that we need to make one time, and if the call fails to work we can just SFTP the file to the proper directory on the server and go directly to the verification page.

So, the method in SiteAdminController is set to use the `.collect` body method with a maxSize of 1gb. If it doesn't work we can copy the file up manually.