Strict Concurrency fixes
This updates the codebase to compile with Swift 6 strict concurrency checks.

Lots of things made sendable. Models are now @unchecked Sendable, per https://blog.vapor.codes/posts/fluent-models-and-sendable/.

WebSocketStore is now an actor, meaning access to its methods is now async. Settings should probably also become an actor, eventually, but many of the places where it's used are in non-async contexts, and we'd have to make a whole bunch of struct initializers async.

Route methods where we were previously using `async let` to parallelize the counting and paged reading of a table, we now do those calls one after the other, as QueryBuilder isn't sendable.

Also, UserController now has a fix for bug #337.