#356 (#365)
* Fixes #360.

Failable initializers are weird.

* Fixes #356.

Adds a Markdown file for the cruise FAQ. And, there is now a common method for displaying styled Markdown in a page with the site header. The cruise FAQ, About Twitarr, and Code Of Conduct pages all use this method. We previously had checked in a Markdown version of the CoC, but weren't using it to render the page.

Previously, the CoC render parsed the oddly-formatted JSON file; I believe "codeofconduct.json" is no longer used by Swiftarr. Keeping the file in place as it's in /public and clients may still try to access it. However, the JSON file should be considered deprecated and we should try to remove it in the future.

---------

Co-authored-by: Chall Fry <chall@challfry.com>