[feature] Puzzle Hunts (#344)
* Add models for Hunt

* Add separate migration for CallInResult enum

* Unlock time should be a datetime

* Form admins can dump JSON into to create a hunt

* Add handler to list all hunts

* Page to view the puzzles in a hunt

with your answers if you're logged in and you have any

* Handler for puzzle detail

body of the puzzle and the ability to see previous callins and call in the answer if you're logged in.

* Sort PuzzleCallIns by creation time

* Answer submissions

* Encode post as plaintext

not a json string

* Fix rendering of submissions

* Remove description from hunt list

since I made it markdown on the hunt page

* Highlight solved puzzles on hunt page

* Use count on arrays in templates

* Add new classes to .jazzy.yaml

* Fix some comments

* Change "Hunt" to "Puzzle Hunt"  in UI

* Basic test hunt

* Make hunt management routes accessible to twitarrteam

instead of admin only.

* Use staticTime for next unlock time

instead of localTime.

* Remove hunt routes from admin group

They were already added to twitarrteam group

* List puzzle hunts on admin page

* Add ability to delete hunts to admin page

* Add admin hunt edit page

Doesn't yet let you edit any part of the hunt, but shows all data ignoring unlock times.

* Allow editing puzzle bodies

* Sort puzzles on edit page by unlock time

* Allow setting or clearing unlockTime

* Allow changing puzzle title

* Allow editing hunt description

* Allow editing hunt title

* Make site routes consistent

Things are plural if they show a list and singular of they show, create, or modify a single. API routes are unaffected because they're named by category.

* Allow changing the answer of a puzzle

If the new answer is different after normalization, retroactively make any existing callin of the new answer correct.

* Use staticTime for callin history table

* Add border to solved puzzles

The subtle success background made them run together.

* Allow editing and adding hints

* Make titles and answer required

* Fix id attributes in hint editing

* Make all hint inputs form-control

* "Puzzle Title" on each puzzle card

* Also normalize away punctuation

* Revert "Add border to solved puzzles"

This reverts commit f9a77194f0ae1616567e5ba81546f5cf38999201.
Turns out it doesn't look good if you have a solved puzzle below an unsolved one because bootstrap doesn't put borders on all sides.

* Move next puzzle unlock out of list

Don't render list at all if no puzzles are unlocked.

* Use tabs consistently in html

* Don't reset form if we're just going to reload

Seems to fix a Firefox-ism where it loads the new content but shows the form as being reset.

* Meta-refresh the hunt page when a puzzle unlocks

Doesn't work if you change the unlock time out from under it

* Add comments for api params/return values

---------

Co-authored-by: Grant Cohoe <grant@grantcohoe.com>