2025 Games list, plus a few fixes
JoCoGamesCatalog.json now sources from the 2025 games list.

The 2025 games list includes BGG links for most of the games on the list, which significantly reduced the amount of data-massaging I had to do to get BGG to find games by the listed name. But, this means that many of the typos in the list are still there--"Heir of Fankenstein", and "Sherlock Homes"--and the code was set to prefer these names over the ones returned by BGG. So, the code now prefers the names BGG returns.