2025 game catalog (#394)
* Fixes #360.

Failable initializers are weird.

* Fixes #356.

Adds a Markdown file for the cruise FAQ. And, there is now a common method for displaying styled Markdown in a page with the site header. The cruise FAQ, About Twitarr, and Code Of Conduct pages all use this method. We previously had checked in a Markdown version of the CoC, but weren't using it to render the page.

Previously, the CoC render parsed the oddly-formatted JSON file; I believe "codeofconduct.json" is no longer used by Swiftarr. Keeping the file in place as it's in /public and clients may still try to access it. However, the JSON file should be considered deprecated and we should try to remove it in the future.

* Fixes #345.

Fixes a few migration reverts.

* 2025 Song Catalog

~600 more songs, same converting from CSV and legacy Mac line endings.

* 2025 Games list, plus a few fixes

JoCoGamesCatalog.json now sources from the 2025 games list.

The 2025 games list includes BGG links for most of the games on the list, which significantly reduced the amount of data-massaging I had to do to get BGG to find games by the listed name. But, this means that many of the typos in the list are still there--"Heir of Fankenstein", and "Sherlock Homes"--and the code was set to prefer these names over the ones returned by BGG. So, the code now prefers the names BGG returns.

---------

Co-authored-by: Chall Fry <chall@challfry.com>