Fixes issue with new text file import for performer-event linking
This code needs to read in the weekday and time from the tab-delimited source and convert it to a Date in the port TZ, since that's the way schedule events are stored.

Also updated for the 2025 TZ changes, and made the call to get the TZ table available for pre-reg.