Fixes #421: Shutternaut Scheduling
There are several new UI elements that only appear for Shutternaut members, giving Shutternauts a way to coordinate the events they'll be covering.

Events now have a button for Shutternauts to self-assign as event photographers. This toggle works in a way similar to the Follow button--the event shows up on the user's Day Planner, they get notifications when the event is about to start, and they can filter events on the Schedule page to only show events they're photographing.

The event view also has a section that shows members of the Shutternaut group which Shutternauts have self-assigned to photograph the event.

For Shutternaut Managers, there is a new toggle button in Event views that lets the manager mark the event as needing to be covered by a photographer. There is also a new filter menu on the Events page, just for Shutternauts, that includes options for showing only events that need to be photographed.

There is also a shared Day Planner view that shows all the events that any Shutternaut is going to be photographing.

This change adds more query options to `/api/v3/events`, adds backwards-compatible fields to the exported user archive, adds a sub-struct to `EventData` for Shutternauts only, and contains database migrations.