Fixes #402.
Eager loading in Fluent has an option to include soft-deleted rows in the eager loaded table. This prevents the Fluent error, but we still don't want to return info about the (soft) deleted Event in this case. So, the PhotostreamImageData initializer now filters these out, returning the generic Location of "On Boat" instead.

I feel that Fluent's OptionalParentEagerLoader maybe shouldn't be throwing errors when it can't find a parent model; this could also be solved by lifting the eager loader code and modifying it to not throw for this case.